requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.0
pydantic==2.9.2
google-generativeai==0.8.3      # supports gemini-2.5-* family
python-dotenv==1.0.1